+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Unreleased
 
 - [#77](https://github.com/gordalina/cachetool/issues/77) Add IPv6 support
-- [#156](https://github.com/gordalina/cachetool/pull/156) Add `--config` argument & yaml extension
+- [#156](https://github.com/gordalina/cachetool/pull/156) Add `--config` argument & yaml extension (@rayderua)
+- [#151](https://github.com/gordalina/cachetool/pull/152) Generate unique identifiers with more entropy (@jonhattan)
+- [#146](https://github.com/gordalina/cachetool/pull/146) Add Eval & File commands
 
 # 5.0.0