+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Unreleased
+# 5.1.0-alpha.1
 
 - [#77](https://github.com/gordalina/cachetool/issues/77) Add IPv6 support
 - [#156](https://github.com/gordalina/cachetool/pull/156) Add `--config` argument & yaml extension (@rayderua)